+++ docs/core/examples/app1.go
@@ -85,7 +85,7 @@ func (msg MsgSend) GetSignBytes() []byte {
 	if err != nil {
 		panic(err)
 	}
-	return bz
+	return sdk.MustSortJSON(bz)
 }
 
 // Implements Msg. Return the signer.
