+++ examples/democoin/x/oracle/types.go
@@ -18,7 +18,7 @@ func (msg Msg) GetSignBytes() []byte {
 	if err != nil {
 		panic(err)
 	}
-	return bz
+	return sdk.MustSortJSON(bz)
 }
 
 // GetSigners implements sdk.Msg
