+++ x/gov/msgs.go
@@ -3,7 +3,6 @@ package gov
 import (
 	"fmt"
 
-	"github.com/cosmos/cosmos-sdk/server"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
@@ -83,11 +82,7 @@ func (msg MsgSubmitProposal) GetSignBytes() []byte {
 	if err != nil {
 		panic(err)
 	}
-	b, err = server.SortJSON(b)
-	if err != nil {
-		panic(err)
-	}
-	return b
+	return sdk.MustSortJSON(b)
 }
 
 // Implements Msg.
@@ -154,11 +149,7 @@ func (msg MsgDeposit) GetSignBytes() []byte {
 	if err != nil {
 		panic(err)
 	}
-	b, err = server.SortJSON(b)
-	if err != nil {
-		panic(err)
-	}
-	return b
+	return sdk.MustSortJSON(b)
 }
 
 // Implements Msg.
@@ -222,11 +213,7 @@ func (msg MsgVote) GetSignBytes() []byte {
 	if err != nil {
 		panic(err)
 	}
-	b, err = server.SortJSON(b)
-	if err != nil {
-		panic(err)
-	}
-	return b
+	return sdk.MustSortJSON(b)
 }
 
 // Implements Msg.
